refactor:BlackJack에서 view 의존
BlackJackView로 분리
