+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -8,22 +8,6 @@ import blackjack.domain.PlayerResult
 import blackjack.domain.PlayerResults
 
 object OutputView {
-    fun showInitialCards(game: BlackJackGame) {
-        println(printProvidedCard(game))
-        println(printDealerCard(game.dealer))
-        game.players.forEach { player ->
-            printPlayerCards(player)
-        }
-    }
-
-    fun printPlayerCards(player: Player) {
-        println(printPlayerCard(player))
-    }
-
-    fun printDealerHaveAdditionalCard() {
-        println(DEALER_GET_ADDITIONAL_CARD)
-    }
-
     fun printFinalCards(game: BlackJackGame) {
         println(printDealerCard(game.dealer) + printTotalSum(game.dealer))
 
@@ -45,10 +29,7 @@ object OutputView {
         }
     }
 
-    private const val DEALER_GET_ADDITIONAL_CARD = "\n딜러는 16이하라 한장의 카드를 더 받았습니다.\n"
     private const val FINAL_RESULT_NOTICE = "\n##최종 승패"
-
-    private fun printProvidedCard(game: BlackJackGame): String = "딜러와 ${game.players.joinToString { it.name }}에게 2장을 나누었습니다.\n"
 
     private fun printDealerCard(dealer: Dealer): String = "딜러: ${dealer.cards.getCards().first().format()}"
